requirements.txt
asgiref==3.6.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-pagination==1.0.7
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
pip==21.3.1
pypng==0.20220715.0
pytz==2021.3
qrcode==7.4.2
reportlab==3.5.59
setuptools==60.2.0
sqlparse==0.4.3
typing-extensions==4.5.0
wheel==0.37.1
djangorestframework-simplejwt
mysqlclient